refactor: HomeTopAppBar 분리
